backend/requirements.txt
Flask
Flask_SQLAlchemy
Flask_Cors
pymysql
python-dotenv
PyJWT
cryptography
bcrypt